requirements.txt
azure-common==1.1.25
azure-core==1.8.0
azure-cosmosdb-nspkg==2.0.2
azure-cosmosdb-table==1.0.6
azure-functions==1.3.0
azure-nspkg==3.0.2
azure-storage-blob==12.4.0
Babel==2.8.0
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
idna==2.10
isodate==0.6.0
Jinja2==2.11.2
MarkupSafe==1.1.1
msrest==0.6.18
oauthlib==3.1.0
pycparser==2.20
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
urllib3==1.25.10
